Docker.ContainerToNodeMetadata: remove needless mocking
Previously it mocked the container and the StateToStatus. It just 
asserted that getter methods were called, rather than confirming the
resulting NodeMetadata had the right values.

By removing some of the mocks, it not only simplifies the code but 
improves code-coverage. (e.g. if StateToStatus was being passed the
wrong value, the previous test would not have noticed).